--- www/logout_completed.php
@@ -1,11 +1,0 @@
-<?php
-
-/**
- * This is the handler for logout completed from the consent page.
- *
- * @package SimpleSAMLphp
- */
-
-$globalConfig = \SimpleSAML\Configuration::getInstance();
-$t = new \SimpleSAML\XHTML\Template($globalConfig, 'consent:logout_completed.twig');
-$t->send();
